requirements.txt
playwright==1.52.0
python-telegram-bot==21.11
python-dotenv==1.1.0
pyyaml==6.0.2
pytest==8.3.5
pytest-asyncio==0.25.3
